CHAM-77 feat : 소켓 브로드캐스트를 위한 훅 작성
