+++ lib/api/axios.ts
@@ -213,3 +213,5 @@ authenticatedApiClient.interceptors.response.use(
     return Promise.reject(error);
   }
 );
+
+export { getAccessToken };
